requirements.txt
langchain-community
langchain-ollama
psycopg2-binary
datasets
transformers
deepeval